Add meshsets for vertices and fill rootset
Add a corresponding meshset for each vertex, and use that in
parent/child relationships.  (Previously, the code was incorrectly
attempting to add a vertex as a child of edge meshsets.)  Also,
add all entities to the root set (or "gather set" in the terminology
of the Cubit plug-in code).

This makes the output much closer to output from an equivalent
Cubit process.  Handling of materials is not yet the same.  Also,
there is an intermittent problem with the faceting, with Open
Cascade sometimes failing to produce edge facets (but then possibly
succeeding upon running exactly the same thing a second time).