Make test output closer to Cubit output
The test compares the number of groups of a certain shape in Cubit
and occ_faceter outputs for similar geometry.  In order to make the
Cubit and occ_faceter test output more similar, I have added a
call to mbtool.gather_ents() in the test (as is already done in the
occ_faceter executable), and I have changed the test data to have a
single material for both volumes.

The output from occ_faceter and Cubit is still different, because
the occ_faceter output includes a material group which contains
two volumes, and the Cubit output instead includes an empty group.
(I don't know what materials groups existed in the Cubit input.)