WIP: Fix node/vertex confusion
This is WIP, and is currently broken.

For every triangle (facet) vertex, a node and meshset were being
added to MOAB.  With this change, a meshset is only added for
vertices which exist in the original geometry, and "nodes" continue
to be added for every triangle (facet) vertex.

Also, a new cylinder faceting test is added to test this change,
since in the other test there are no nodes which are not vertices,
so this change should make no difference there.

This commit also includes a bug fix in raise_moab_error(), where
unexpected rvals were causing a dodgy memory access.