Fix test output to match Cubit output
The test output that we have from Cubit only matches occ_faceter
output for one empty material group.  We should try to recreate
the test output from Cubit with more suitable material group input,
but I didn't have time to do that now, so this is a quick fix so
that the current test passes and can be used to confirm that the
rest of the behaviour remains as we expect.

We already know that work is needed on propogating materials, and
this hack should be fixed as part of that work.